Added http request for adding, editing and deleting Todos. Also refactored repeated code
