--- src/utils/ErrorMessage.jsx
@@ -1,11 +1,0 @@
-const ErrorMessage = ({ errorMsg, dimissErrorHandler }) => {
-  return (
-     <div>
-      <hr />
-      <p>{errorMsg}</p>
-      <button type='button' onClick={dimissErrorHandler}>Dismiss</button>
-    </div> 
-  )
-}
-
-export default ErrorMessage;
