+++ src/utils/recordMapper.js
@@ -1,0 +1,10 @@
+export const recordMapper = (record) => {
+  const todo = {
+    id: record.id,
+    ...record.fields,
+  }
+  if (!record.fields.isCompleted) {
+    todo.isCompleted = false;
+  }
+  return todo;
+}
